+++ app/data_access/mission.py
@@ -1,29 +1,15 @@
 import graphene
 from graphene.types.generic import GenericScalar
 
+from app.helpers.frozen_version_utils import retrieve_max_reception_time
 from app.helpers.graphene_types import BaseSQLAlchemyObjectType, TimeStamp
 from app.models import Mission
 from app.data_access.activity import ActivityOutput
 from app.helpers.authentication import current_user
 from app.models.comment import CommentOutput
-from app.models.controller_control import ControllerControl
 from app.models.expenditure import ExpenditureOutput
-from app.models.location_entry import LocationEntryType, LocationEntryOutput
+from app.models.location_entry import LocationEntryOutput
 from app.models.mission_validation import MissionValidationOutput
-
-
-def retrieve_max_reception_time(info):
-    # TODO à discuter en review
-    # if info.context.view_args["max_reception_time"]:
-    if (
-        info.path[0] == "controlData"
-        or info.path[0] == "readMissionControlData"
-    ):
-        controller_control = ControllerControl.query.get(
-            info.variable_values["controlId"]
-        )
-        return controller_control.qr_code_generation_time
-    return None
 
 
 class MissionOutput(BaseSQLAlchemyObjectType):
